+++ lib/v2/mihoyo/bbs/img-ranking.js
@@ -1,10 +1,6 @@
 const got = require('@/utils/got');
-const { art } = require('@/utils/render');
-const path = require('path');
-const { parseDate } = require('@/utils/parse-date');
 const { DATA_MAP, RANKING_TYPE_MAP } = require('./static-data');
-
-const renderDescription = (description, images) => art(path.join(__dirname, '../templates/description.art'), { description, images });
+const { post2item } = require('./utils');
 
 const getGameInfo = (game) => ({
     gids: DATA_MAP[game]?.gids,
@@ -75,32 +71,7 @@ module.exports = async (ctx) => {
         throw new Error('未获取到数据！');
     }
     const title = `米游社-${game_title}-${forum_title}${cate_title ? `-${cate_title}` : ''}-${type_title}`;
-    const items = list.map((e) => {
-        const author = e.user.nickname;
-        const title = e.post.subject;
-        const link = `https://www.miyoushe.com/ys/article/${e.post.post_id}`;
-        let describe = e.post.content || '';
-        try {
-            describe = JSON.parse(e.post.content).describe;
-        } catch (error) {
-            if (!(error instanceof SyntaxError)) {
-                throw error;
-            }
-        }
-        const description = renderDescription(describe || '', [...new Set([e.post.cover, ...e.post.images])].filter(Boolean));
-        const pubDate = parseDate(e.post.created_at * 1000);
-        const upvotes = e.stat.like_num;
-        const comments = e.stat.reply_num;
-        return {
-            author,
-            title,
-            link,
-            description,
-            pubDate,
-            upvotes,
-            comments,
-        };
-    });
+    const items = list.map((e) => post2item(e));
     const data = {
         title,
         link: url,
