+++ lib/v2/mihoyo/bbs/official.js
@@ -1,7 +1,5 @@
 const got = require('@/utils/got');
-const { art } = require('@/utils/render');
-const path = require('path');
-const { parseDate } = require('@/utils/parse-date');
+const { post2item } = require('./utils');
 // 游戏id
 const GITS_MAP = {
     1: '崩坏三',
@@ -18,8 +16,6 @@ const TYPE_MAP = {
     2: '活动',
     3: '资讯',
 };
-
-const renderDescription = (description, images) => art(path.join(__dirname, '../templates/description.art'), { description, images });
 
 module.exports = async (ctx) => {
     const { gids, type = '2', page_size = '20', last_id = '' } = ctx.params;
@@ -39,20 +35,7 @@ module.exports = async (ctx) => {
         throw new Error('未获取到数据！');
     }
     const title = `米游社 - ${GITS_MAP[gids] || ''} - ${TYPE_MAP[type] || ''}`;
-    const items = list.map((e) => {
-        const author = e.user.nickname;
-        const title = e.post.subject;
-        const link = `https://www.miyoushe.com/ys/article/${e.post.post_id}`;
-        const description = renderDescription(e.post.content, e.post.images);
-        const pubDate = parseDate(e.post.created_at * 1000);
-        return {
-            author,
-            title,
-            link,
-            description,
-            pubDate,
-        };
-    });
+    const items = list.map((e) => post2item(e));
     const data = {
         title,
         link: url,
