+++ lib/v2/mihoyo/bbs/timeline.js
@@ -1,11 +1,7 @@
 const got = require('@/utils/got');
-const { art } = require('@/utils/render');
-const path = require('path');
-const { parseDate } = require('@/utils/parse-date');
 const cache = require('./cache');
 const config = require('@/config').value;
-
-const renderDescription = (description, images) => art(path.join(__dirname, '../templates/description.art'), { description, images });
+const { post2item } = require('./utils');
 
 module.exports = async (ctx) => {
     if (!config.mihoyo.cookie) {
@@ -34,32 +30,7 @@ module.exports = async (ctx) => {
     }
     const { nickname: username } = await cache.getUserFullInfo(ctx, '');
     const title = `米游社 - ${username} 的关注动态`;
-    const items = list.map((e) => {
-        const author = e.user.nickname;
-        const title = e.post.subject;
-        const link = `https://www.miyoushe.com/ys/article/${e.post.post_id}`;
-        let describe = e.post.content || '';
-        try {
-            describe = JSON.parse(e.post.content).describe;
-        } catch (error) {
-            if (!(error instanceof SyntaxError)) {
-                throw error;
-            }
-        }
-        const description = renderDescription(describe || '', [...new Set([e.post.cover, ...e.post.images])].filter(Boolean));
-        const pubDate = parseDate(e.post.created_at * 1000);
-        const upvotes = e.stat.like_num;
-        const comments = e.stat.reply_num;
-        return {
-            author,
-            title,
-            link,
-            description,
-            pubDate,
-            upvotes,
-            comments,
-        };
-    });
+    const items = list.map((e) => post2item(e));
 
     const data = {
         title,
